Move shimmer into :common:ui module
